Split match modules based on OF version Updated coveragerc to ignore OF1.0
