+++ .coveragerc
@@ -1,3 +1,3 @@
 [run]
 source = .
-omit = .eggs/*,.tox/*,*tests*,setup.py
+omit = .eggs/*,.tox/*,*tests*,setup.py,.direnv/*,.venv/*,venv/*,v0x01/match.py
