+++ match.py
@@ -1,10 +1,8 @@
 """Switch match."""
 
-import ipaddress
-
-from pyof.v0x01.common.flow_match import FlowWildCards
-
-IPV4_ETH_TYPE = 2048
+from napps.kytos.flow_manager.v0x01.match import match10_no_strict
+from napps.kytos.flow_manager.v0x04.match import match13_no_strict
+from napps.kytos.flow_manager.v0x04.match import match13_strict
 
 
 def match_flow(flow_to_install, version, stored_flow_dict):
@@ -30,157 +28,3 @@ def match_strict_flow(flow_to_install, version, stored_flow_dict) -> None:
     if version != 0x04:
         raise NotImplementedError(f"Unsupported OpenFlow version {version}")
     return match13_strict(flow_to_install, stored_flow_dict)
-
-
-def _get_match_fields(flow_dict):
-    """Generate match fields."""
-    match_fields = {}
-    if "match" in flow_dict:
-        for key, value in flow_dict["match"].items():
-            match_fields[key] = value
-    return match_fields
-
-
-# pylint: disable=too-many-return-statements, too-many-statements, R0912
-def _match_ipv4_10(match_fields, args, wildcards):
-    """Match IPV4 fields against packet with Flow (OF1.0)."""
-    if match_fields.get("dl_type") == IPV4_ETH_TYPE:
-        return False
-    flow_ip_int = int(ipaddress.IPv4Address(match_fields.get("nw_src", 0)))
-    if flow_ip_int != 0:
-        mask = (
-            wildcards & FlowWildCards.OFPFW_NW_SRC_MASK
-        ) >> FlowWildCards.OFPFW_NW_SRC_SHIFT
-        mask = min(mask, 32)
-        if mask != 32 and "nw_src" not in args:
-            return False
-        mask = (0xFFFFFFFF << mask) & 0xFFFFFFFF
-        ip_int = int(ipaddress.IPv4Address(args.get("nw_src")))
-        if ip_int & mask != flow_ip_int & mask:
-            return False
-    flow_ip_int = int(ipaddress.IPv4Address(match_fields.get("nw_dst", 0)))
-    if flow_ip_int != 0:
-        mask = (
-            wildcards & FlowWildCards.OFPFW_NW_DST_MASK
-        ) >> FlowWildCards.OFPFW_NW_DST_SHIFT
-        mask = min(mask, 32)
-        if mask != 32 and "nw_dst" not in args:
-            return False
-        mask = (0xFFFFFFFF << mask) & 0xFFFFFFFF
-        ip_int = int(ipaddress.IPv4Address(args.get("nw_dst")))
-        if ip_int & mask != flow_ip_int & mask:
-            return False
-    if not wildcards & FlowWildCards.OFPFW_NW_TOS:
-        if ("nw_tos", "nw_proto", "tp_src", "tp_dst") not in args:
-            return True
-        if match_fields.get("nw_tos") != int(args.get("nw_tos")):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_NW_PROTO:
-        if match_fields.get("nw_proto") != int(args.get("nw_proto")):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_TP_SRC:
-        if match_fields.get("tp_src") != int(args.get("tp_src")):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_TP_DST:
-        if match_fields.get("tp_dst") != int(args.get("tp_dst")):
-            return False
-    return True
-
-
-# pylint: disable=too-many-return-statements, too-many-statements, R0912
-def match10_no_strict(flow_dict, args):
-    """Match a packet against this flow (OF1.0)."""
-    args = _get_match_fields(args)
-    match_fields = _get_match_fields(flow_dict)
-    wildcards = match_fields.get("wildcards", 0)
-    if not wildcards & FlowWildCards.OFPFW_IN_PORT:
-        if match_fields.get("in_port") != args.get("in_port"):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_DL_VLAN_PCP:
-        if match_fields.get("dl_vlan_pcp") != args.get("dl_vlan_pcp"):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_DL_VLAN:
-        if match_fields.get("dl_vlan") != args.get("dl_vlan"):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_DL_SRC:
-        if match_fields.get("dl_src") != args.get("dl_src"):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_DL_DST:
-        if match_fields.get("dl_dst") != args.get("dl_dst"):
-            return False
-    if not wildcards & FlowWildCards.OFPFW_DL_TYPE:
-        if match_fields.get("dl_type") != args.get("dl_type"):
-            return False
-    if not _match_ipv4_10(match_fields, args, wildcards):
-        return False
-    return flow_dict
-
-
-def _match_cookie(flow_to_install, stored_flow_dict):
-    """Check if a the cookie and its mask matches between the flows."""
-    cookie = flow_to_install.get("cookie", 0) & flow_to_install.get("cookie_mask", 0)
-    cookie_stored = stored_flow_dict.get("cookie", 0) & flow_to_install.get(
-        "cookie_mask", 0
-    )
-    if cookie and cookie != cookie_stored:
-        return False
-    return True
-
-
-def _match_keys(flow_to_install, stored_flow_dict, flow_to_install_keys):
-    """Check if certain keys on flow_to_install match on stored_flow_dict."""
-    for key in flow_to_install_keys:
-        if key not in stored_flow_dict["match"]:
-            return False
-        if flow_to_install["match"][key] != stored_flow_dict["match"].get(key):
-            return False
-    return True
-
-
-def match13_no_strict(flow_to_install, stored_flow_dict):
-    """Match a flow that is either exact or more specific (non-strict) (OF1.3).
-
-    Return the flow if any fields match, otherwise, return False.
-    """
-    if not _match_cookie(flow_to_install, stored_flow_dict):
-        return False
-
-    if "match" not in flow_to_install or "match" not in stored_flow_dict:
-        return stored_flow_dict
-    if not flow_to_install["match"]:
-        return stored_flow_dict
-    if len(flow_to_install["match"]) > len(stored_flow_dict["match"]):
-        return False
-
-    if not _match_keys(
-        flow_to_install, stored_flow_dict, flow_to_install["match"].keys()
-    ):
-        return False
-    return stored_flow_dict
-
-
-def match13_strict(flow_to_install, stored_flow_dict):
-    """Match a flow strictly (OF1.3).
-
-    Return the flow if all fields match, otherwise, return False.
-    """
-    if not _match_cookie(flow_to_install, stored_flow_dict):
-        return False
-    if flow_to_install.get("priority", 0) != stored_flow_dict.get("priority", 0):
-        return False
-
-    if "match" not in flow_to_install and "match" not in stored_flow_dict:
-        return stored_flow_dict
-    if "match" not in flow_to_install and "match" in stored_flow_dict:
-        return False
-    if "match" in flow_to_install and "match" not in stored_flow_dict:
-        return False
-
-    if len(flow_to_install["match"]) != len(stored_flow_dict["match"]):
-        return False
-
-    if not _match_keys(
-        flow_to_install, stored_flow_dict, flow_to_install["match"].keys()
-    ):
-        return False
-    return stored_flow_dict
